starter_py/requirements.txt
django==3.1
python-decouple==3.3
djangorestframework==3.11.1
firebase-admin==4.3.0
